refactored Gui mod
